s3ext: refactor gpreader
1. write the GPReader class to connect all readers

2. add unit tests on Open, Construct, Close, Read,
   multi-thread Read, and failed response handling

3. and a few cleanup
  3.1. clean source codes in s3ext
  3.2. refactor debug switch method
  3.3. move mockRESTfulService to mock_classes.h
  3.4. reset QueryCancelPending to FALSE after tests

+++ gpAux/extensions/gps3ext/include/reader_params.h
@@ -90,9 +90,9 @@ class ReaderParams {
     string urlToLoad;  // original url to read/write.
     string keyUrl;     // key url in s3 bucket.
     string region;
-    uint64_t keySize;     // key/file size.
-    uint64_t chunkSize;   // chunk size
-    uint8_t numOfChunks;  // number of chunks(threads).
+    uint64_t keySize;      // key/file size.
+    uint64_t chunkSize;    // chunk size
+    uint64_t numOfChunks;  // number of chunks(threads).
     S3Credential cred;
     int segId;
     int segNum;
